Refactor TSE4000 optimizer modules
