+++ trading/tse/__init__.py
@@ -1,0 +1,16 @@
+"""Tokyo Stock Exchange optimization toolkit."""
+
+from .analysis import StockAnalyzer, StockProfile
+from .optimizer import PortfolioOptimizer, PORTFOLIO_SIZES, DEFAULT_TARGET_SIZE
+from .backtester import PortfolioBacktester
+from .reporting import OptimizationReporter
+
+__all__ = [
+    "StockAnalyzer",
+    "StockProfile",
+    "PortfolioOptimizer",
+    "PORTFOLIO_SIZES",
+    "DEFAULT_TARGET_SIZE",
+    "PortfolioBacktester",
+    "OptimizationReporter",
+]
